pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-evernote"
version = "0.1.0"
description = "Save Claude AI chats as Evernote notes"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "evernote3>=1.25.14",
    "markdown>=3.5",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.scripts]
claude-evernote = "claude_evernote.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/claude-evernote-connector"
Issues = "https://github.com/yourusername/claude-evernote-connector/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["claude_evernote*"]
